Tested updated get_feature_importances
