+++ cqu/classical/__init__.py
@@ -28,13 +28,4 @@ Typical usage example:
 
 """
 
-from enum import Enum
-
-
-class ClassicalModels(Enum):
-    LOGISTIC_REGRESSION = "logistic_regression"
-    RANDOM_FOREST = "random_forest"
-    GRADIENT_BOOSTING = "gradient_boosting"
-    NEURAL_NETWORK = "neural_network"
-    KNN = "knn"
-    NAIVE_BAYES = "naive_bayes"
+from .models import ClassicalModels
